add phasechange test stub
