+++ pkg/controllers/dynakube/controller_integration_test.go
@@ -19,7 +19,9 @@ import (
 	"github.com/Dynatrace/dynatrace-operator/pkg/controllers/dynakube/istio"
 	"github.com/Dynatrace/dynatrace-operator/pkg/controllers/dynakube/version"
 	"github.com/Dynatrace/dynatrace-operator/pkg/util/kubeobjects/address"
+	"github.com/Dynatrace/dynatrace-operator/pkg/util/kubeobjects/labels"
 	"github.com/Dynatrace/dynatrace-operator/pkg/util/kubesystem"
+	version2 "github.com/Dynatrace/dynatrace-operator/pkg/version"
 	"github.com/spf13/afero"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -29,6 +31,7 @@ import (
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/util/intstr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 )
@@ -599,4 +602,155 @@ func TestAPIError(t *testing.T) {
 		assert.NoError(t, err)
 		assert.Equal(t, fastUpdateInterval, result.RequeueAfter)
 	})
+}
+
+// generateStatefulSetForTesting prepares an ActiveGate StatefulSet after a Reconciliation of the Dynakube with a specific feature enabled
+func generateStatefulSetForTesting(name, namespace, feature, kubeSystemUUID string) *appsv1.StatefulSet {
+	expectedLabels := map[string]string{
+		labels.AppNameLabel:      labels.ActiveGateComponentLabel,
+		labels.AppVersionLabel:   testComponentVersion,
+		labels.AppComponentLabel: feature,
+		labels.AppCreatedByLabel: name,
+		labels.AppManagedByLabel: version2.AppName,
+	}
+	expectedMatchLabels := map[string]string{
+		labels.AppNameLabel:      labels.ActiveGateComponentLabel,
+		labels.AppManagedByLabel: version2.AppName,
+		labels.AppCreatedByLabel: name,
+	}
+	return &appsv1.StatefulSet{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      name + "-" + feature,
+			Namespace: namespace,
+			Labels:    expectedLabels,
+			OwnerReferences: []metav1.OwnerReference{
+				{
+					APIVersion: "dynatrace.com/v1beta1",
+					Kind:       "DynaKube",
+					Name:       name,
+				},
+			},
+		},
+		Spec: appsv1.StatefulSetSpec{
+			Selector: &metav1.LabelSelector{
+				MatchLabels: expectedMatchLabels,
+			},
+			Template: corev1.PodTemplateSpec{
+				ObjectMeta: metav1.ObjectMeta{
+					Labels: expectedLabels,
+					Annotations: map[string]string{
+						"internal.operator.dynatrace.com/custom-properties-hash": "",
+						"internal.operator.dynatrace.com/version":                "",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Volumes: []corev1.Volume{
+						{
+							Name: "truststore-volume",
+						},
+					},
+					InitContainers: []corev1.Container{
+						{
+							Name: "certificate-loader",
+							Command: []string{
+								"/bin/bash",
+							},
+							Args: []string{
+								"-c",
+								"/opt/dynatrace/gateway/k8scrt2jks.sh",
+							},
+							WorkingDir: "/var/lib/dynatrace/gateway",
+							VolumeMounts: []corev1.VolumeMount{
+								{
+									Name:             "truststore-volume",
+									MountPath:        "/var/lib/dynatrace/gateway/ssl",
+									MountPropagation: (*corev1.MountPropagationMode)(nil),
+								},
+							},
+							ImagePullPolicy: "Always",
+						},
+					},
+					Containers: []corev1.Container{
+						{
+							Name: feature,
+							Env: []corev1.EnvVar{
+								{
+									Name:  "DT_CAPABILITIES",
+									Value: "kubernetes_monitoring",
+								},
+								{
+									Name:  "DT_ID_SEED_NAMESPACE",
+									Value: namespace,
+								},
+								{
+									Name:  "DT_ID_SEED_K8S_CLUSTER_ID",
+									Value: kubeSystemUUID,
+								},
+								{
+									Name:  "DT_DEPLOYMENT_METADATA",
+									Value: "orchestration_tech=Operator-active_gate;script_version=snapshot;orchestrator_id=" + kubeSystemUUID,
+								},
+							},
+							VolumeMounts: []corev1.VolumeMount{
+								{
+									Name:      "truststore-volume",
+									ReadOnly:  true,
+									MountPath: "/opt/dynatrace/gateway/jre/lib/security/cacerts",
+									SubPath:   "k8s-local.jks",
+								},
+							},
+							ReadinessProbe: &corev1.Probe{
+								ProbeHandler: corev1.ProbeHandler{
+									HTTPGet: &corev1.HTTPGetAction{
+										Path: "/rest/health",
+										Port: intstr.IntOrString{
+											IntVal: 9999,
+										},
+										Scheme: "HTTPS",
+									},
+								},
+								InitialDelaySeconds: 90,
+								PeriodSeconds:       15,
+								FailureThreshold:    3,
+							},
+							ImagePullPolicy: "Always",
+						},
+					},
+					ServiceAccountName: "dynatrace-kubernetes-monitoring",
+					ImagePullSecrets: []corev1.LocalObjectReference{
+						{
+							Name: name + "-pull-secret",
+						},
+					},
+					Affinity: &corev1.Affinity{
+						NodeAffinity: &corev1.NodeAffinity{
+							RequiredDuringSchedulingIgnoredDuringExecution: &corev1.NodeSelector{
+								NodeSelectorTerms: []corev1.NodeSelectorTerm{
+									{
+										MatchExpressions: []corev1.NodeSelectorRequirement{
+											{
+												Key:      "kubernetes.io/arch",
+												Operator: "In",
+												Values: []string{
+													"amd64",
+												},
+											},
+											{
+												Key:      "kubernetes.io/os",
+												Operator: "In",
+												Values: []string{
+													"linux",
+												},
+											},
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+			PodManagementPolicy: "Parallel",
+		},
+	}
 }
